+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: abarot <abarot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 17:15:09 by abarot            #+#    #+#             */
-/*   Updated: 2019/10/10 13:14:53 by abarot           ###   ########.fr       */
+/*   Updated: 2019/10/18 16:31:01 by abarot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memset(void *p, int c, size_t n)
 {
-	unsigned int	i;
-	char			*b;
+	size_t		i;
+	char		*b;
 
 	b = (char *)p;
 	i = 0;
